requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pytest>=7.0.0
pytest-asyncio>=0.24.0
httpx>=0.26.0
